backend/requirements.txt
# Core
Django~=5.2.7
djangorestframework~=3.16.1
djangorestframework-simplejwt~=5.5.1
django-cors-headers~=4.9.0
python-decouple==3.8
cryptography~=46.0.1
python-jose~=3.5.0
pydantic~=2.12.4
adrf~=0.1.11


# Database
psycopg2-binary~=2.9.11

# Real Time
channels~=4.3.1
types-channels~=4.3.0.20250822
channels-redis~=4.3.0


# Background Tasks
celery~=5.5.3
redis~=7.0.1

# AI
langchain~=1.0.2
langchain-google-genai~=3.0.0


# Utilities
requests~=2.32.5
pytz==2025.2
drf-spectacular==0.29.0
sentry-sdk~=2.43.0
black~=25.12.0
flake8~=7.3.0


# test
pytest~=9.0.2
pytest-django~=4.11.1
pytest-cov~=7.0.0
factory-boy~=3.3.3
Faker~=38.2.0
freezegun~=1.5.5

# Production
daphne~=4.2.1